still missing some methodes

+++ src/main/java/com/comeback/securityauthback/repository/SchoolClassRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SchoolClassRepo extends JpaRepository<SchoolClass,Long> {
+public interface SchoolClassRepo extends JpaRepository<SchoolClass,Integer> {
 }
